application/backend_entry/backend/requirements.txt
Django==1.11.17
django-environ==0.4.1
djangorestframework==3.5.3
psycopg2-binary==2.7.7
pytz==2018.9
six==1.12.0
elasticsearch
django_saml2_auth
tika
pillow
django-cors-headers
moviepy